WABA: Best innovation since sliced bread: blocking cookies!
When a user is asked whether he/she wants to accept a cookie and
a HTTP slave wants to fetch a page which would use this cookie, the
cookieserver blocks the HTTP slave till the user has made up his/her mind.

This comes in especially handy in those cases where the cookie is set
in a "redirect header" and where the "redirected-to-location" checks
if the cookie is set.

svn path=/trunk/kdebase/kioslave/; revision=37087