* Made it possible to start up the debugger from Dr Konqi. * The debugger page was renamed to backtrace.
If you want to to choose a different debugger, add the settings below to
your drkonqirc file:
[drkonqi]
Debugger=kdbg

And then add a "kdbgrc" file to the debuggers directory. (Just take a look
at gdbrc for the format.)

NOTE: The debugger is turned of by default. If you want it, please change
your drkonqirc file to include:
[drkonqi]
ConfigName=developer

This will be configurable from kcmdrkonqi in the future.

svn path=/trunk/kdebase/drkonqi/; revision=47917