Added support for .anchors files.
(What these actually do is one of the best hidden secrest
in KDE: They link symbolic links inside the documentation
to the actual filenames. This is needed because the traslation
team thinks it is very important to translate the filenames
of the translations)

And yes, the implementation uses another redirection.

svn path=/trunk/kdebase/kioslave/; revision=51171