IMAP4rev1 client in a class. Hopefully now whoever wants to make the IMAP ioslave work will use this. It's a fairly complete implementation.
I need to implement AUTHENTICATE. LOGIN works, so you can send your
passwords in the clear if you don't care about security. Probably the
stuff from the current slave code can be used to write the authenticate()
function.

This could also do with having some parsing of the retval of FETCH,
but it's not really necessary, it would just make it even easier to
use. BTW, it's _really_ easy to use :)

Uses QIODevice, so you can plug it in anywhere.

svn path=/trunk/kdebase/kioslave/; revision=71063