Adding the new ioslave sources that are based on the libsmbclient library from the Samba organization.
This code is pretty stable- there are still a couple of things to work out
with authentication issues, but it works for now for the most part.

I added a small howto on getting the libsmbclient downloaded and built.
It turns out you don't need the whole Samba source, just the samba/source
directory which is only about 10 megs, so people can grab it pretty
easily.

I'll add the autoconf stuff in tomorrow.

svn path=/trunk/kdebase/kioslave/; revision=87063