This was slowly bitrotting on my HD since some months (!). audiocd:/ now finally supports local databases (xmcd format), and also locally saves any title lists it got over the net.  Makes everything more snappy, plus doesn't require a inet connection anymore.
I would appriciate if someone goes over the configuring GUI.  Note, that
internally the files from the net are saved in the _first_ directory in the
list.  All dirs in that list are searched.  So the first one should be
a writable dir (it's created when necessary).  I for instance have
  .cddb/
  /usr/X11R6/lib/X11/xmcd/cddb
in my dirlist.

svn path=/trunk/kdebase/kioslave/; revision=121670