Here it comes:
- moved everything back to designer
- redesigned the preview widget to be less space consuming
- ignore RMB clicks on the preview widgets, too
- beautify the style preview by adding a fake Titlebar (Post 3.0 TODO: move
it to QWorkspace or a yet-to-be-written KWorkspace and make this one
work with kwin-styles, idea by gallium and fred1)
- make almost anything else work again (exept for the X-Render settings)

Gallium: does the acutual save routine really change the styles at runtime?
Here it doesn't work neither for KDE apps nor for Qt apps. Did I break
something or did you forget to commit hacks in kdelibs?

svn path=/trunk/kdebase/kcontrol/style/; revision=133667