CCMAIL: 60885@bugs.kde.org
Try to be smart when detecting the most top level domain to traverse to
when looking for a host named "wpad". The first domain for which a SOA
record exists is the last one tried.
Michal, can you please check if this works for you now?

Don't crash kded if proyForURL() is called but autoproxy turned off.

svn path=/trunk/kdelibs/kio/misc/; revision=236128