ACL support (RFC 2086). Implemented SETACL, DELETEACL, GETACL, and MYRIGHTS, using KIO::special().
I didn't implement LISTRIGHTS, not needed atm (but that's easy to add if needed).
If we provide default sets of permissions to be set (instead of the full 2^9
flexibility), we don't need to find out which permissions are tied together.

Some simple speed fixes. Fixed the famous "bad bad bad" warning.

svn path=/trunk/kdebase/kioslave/; revision=301637