Add support to forward KNotify's popup messages to DBus interface if it exists on session bus. In particular, this will allow a closer interaction of plasma with KNotify's popups.
Supported DBus interface is an interface which is currently a galago-like interface, but with
different name, which can be changed in future when it will be decided if we'll collaborate with
galago guys to bring some improvements into the existing spec, or create something new if the former
will not be possible for some reason.

This patch was reviewed on kde-core-devel and the corresponding feature entry exists on Features page
on a techbase, so I hope it is ok to commit this now -  as I want this to get in before hard feature 
freeze will happen.

If you notice any bugs or will have any suggestions, please mail me - dimsuz at gmail dottie-here com

svn path=/trunk/KDE/kdebase/runtime/; revision=803967