Respect user configured icon effects for menu icons
Revert commit r792492. The code had only been added to work around
bug 165460, which is now fixed.

The additional benefit is that disabled menu icons now go through the
standard pixmap caches.

If the slight change in transparency is a problem, KIconEffect can
be enhanced to allow an opacity parameter.

svn path=/trunk/KDE/kdebase/runtime/; revision=952132