Backport to 4.3branch of: SVN commit 988281 by darioandres:
Show the "Show Contents of the Report" button in the last page of the 
reporting assistant if the submitting fails. This will allow the reporter
to save the report information to submit it manually later.

A further improvement (for 4.4) will be warning the user to save the report
data if he wants to close the assistant without reporting
(or if he couldn't do it)

CCBUG: 198041


svn path=/branches/KDE/4.3/kdebase/runtime/; revision=988348