Try to fix the famous my keyboard is not working bug by providing CurrentTime to XUngrabKeyboard. According to the documentation if not doing that the ungrabbing can silently fail.
I have no idea what that change really does but i ran it the complete day
and noticed no difference.

I had the problem yesterday on my laptop twice. Today it didn't occur.

Please report back if running on trunk and the problem happens for you with
this patch.

CCBUG: 171685

svn path=/trunk/KDE/kdebase/runtime/; revision=993433