Support for Anchor Layout (Qt 4.6)
Now we have support for anchor layout on javascript
plasmoids. This is at 90%. It's not working properly yet
as we need either a patch in Qt (being discussed) or another
not so beautiful approach. After this it will work 98%, and
I'll just need to do the QGraphicsAnchor bindings to support
spacing and size policies of anchors.


svn path=/trunk/KDE/kdebase/runtime/; revision=1044107