phonon: Convert the phonon kcm to use the new GlobalConfig system now exported in libphonon.
Previously, the code here would look directly at the configuration files used
by libphonon in order to tweak the devices and reorder their priority.

This meant that a lot of code was copied from libphonon to here which is not
really ideal from a Good Programming Practice/code reuse perspective.

In addition to this, if the new PulseAudio support in phonon is in use then
the config files themselves are not consulted (all of the priority list
handling being handed over to PA) thus making the adjustment of the
configuration files rather irrelevent.

The solution implemented here is to use the new GlobalConfig class which is
now exported from from libphonon. This class basically wraps up the
configuration and either reads/writes the configuration files (when PA is not
in use) or reads/writes from the module-device-manager extension of PA which
I've specifically written to implement the routing policy used in KDE/phonon.

Note: One functional change is that the ticky box for showing advanced devices
is now written to the config immediately rather than waiting for the user to
hit the Apply button. This is needed so as to retrieve an approriate listing
of the devices from the configuration system (which looks at the advanced flag
to determin what devices to return via a different (existing) API).

Note 2: There is only one direct access to the config files left which relates
to an version update which I think we can just leave or eventually remove
completely (if the loss of these settings on upgrade is acceptable).

This change means that we rely on phonon trunk from kdesupport.

svn path=/trunk/KDE/kdebase/runtime/; revision=1051736