Add ISO Currency Code support to KLocale
See reviewboard http://reviewboard.kde.org/r/2049/

Currency files part, commit 2 of 4 to be done.

These are all current currencies, and historic ones already supported by
KSpread and KMyMoney. There are about 100 more that can be added later if
needed.  Hopefully none missed and not too many spelling errors in the names.
Note not all fields are currently used, but it is planned they will be in the
future.

svn path=/trunk/KDE/kdebase/runtime/; revision=1054364