Create bindings for QGraphicsGridLayout
We should definetely have a common header for the layout bindings
so we avoid duplicating the "layoutItem" function with different
names all around.

I created some macros to help creating the binding for GridLayout
that may be helpful in other scenarious. Maybe put this macros on
this shared header too ?

Anyway, the bindings for this class are ready if I didn't forget
any important functions and the code for layout item was a little
bit simplified trying to qscriptvalue_cast to QGraphicsLayout* instead
of each implementation of it.


svn path=/trunk/KDE/kdebase/runtime/; revision=1055930