Forwardport r1075552:
Fix 100% CPU usage due to a startTimer(0) that was never stopped; this could
happen if the timer was started twice, so m_timerId was overwritten, and the first
timer ID never recognized. It's much much simpler to use deleteLater: no risk of forgetting
the deletion and no risk of timers running forever. It means using QWeakPointer to notice
when it got deleted, though. BUG 202744, possibly also 178228 184576 217364 220047.

svn path=/trunk/KDE/kdebase/runtime/; revision=1075556