Fix finding SDO.
kdebase couldn't be compiled, complaining about missing function/classes in a
namespace. It took me 30 minutes to find out that:
a) the files where this is are generated
b) the source files are from a too-old SDO version

The problem is that the check only lists SDO as an optional package that
couldn't be found, but still tries to build the nepomuk stuff. So either the
nepomuk stuff needs to be disabled if SDO is not found, or SDO needs to be a
requirement. This was just the easier fix for now, hence cc'ing kde-buildsystem
so a proper decision can be done which fix is the right one.

CCMAIL: kde-buildsystem@kde.org


svn path=/trunk/KDE/kdebase/runtime/; revision=1113683