Revert last commit and apply a better fix
The problem was that the module saved an empty exec line
as "!", instead of resetting to KIO based service. The bug
should be catched on saving instead of on loading, otherwise
the user believes she is using KIO while in reality it tries
to run an empty command (and causes a crash this way).

CCBUG: 240677

svn path=/trunk/KDE/kdebase/runtime/; revision=1135569