phonon: Add a GUI to configure and control the sound setup.
This GUI is only available if PulseAudio is used, but it allows you to
configure Stereo vs. 5.1 Surround vs. Digital/IEC958 configuration
of your hardware.

The GUI also is able to send test sounds to your speakers.
It uses libcanberra for this (which is a pretty common counterpart
to a PulseAudio system). The sounds used are part of the Free Desktop
sound theme (i.e. the fallback sound theme - the hicolor of sounds).

This commit should work fine when PulseAudio development headers are not
installed and should work fine at runtime when PulseAudio is not in use
(i.e. it fails gracefully). The tab itself is not presented when
either of those two conditions are met.

See http://colin.guthr.ie/2010/07/speaker-setup-now-or-forever-hold-your-peace/

svn path=/trunk/KDE/kdebase/runtime/; revision=1154776