Add the last few ISO codes for populated places, update the C settings.
Add Jersey, Gurnsey, Isle of Man, Saint Martin, Saint Bartholemy, and Northern
Mariana Islands.

I still need to find proper flags, but I wanted to get in for the i18n freeze.

The C locale now has all the default values in so it can be used to set the
default defaults in both KLocale and the KCM, rather than having to hard code
them in both places and the obvious mistakes that wil lead to.


svn path=/trunk/KDE/kdebase/runtime/; revision=1195673