Improved version of the Locale Control Module.
This is an improved version of the Locale KCM designed to help users see where
KLocale has auto-detected settings from the system, and where they have over-
ridden them.  A small Default Arrow appears next to each setting, showing
when the user has overridden and allowing them to reset just that setting.

The language selector is improved using standard KDE widgets, and a number of
new options are supported.

The ui is now defined in a single ui file instead of code, but I may break it
down in separate ui files per tab in 4.7, but for now it was good to
consolidate the code and make it consistent.

I'll admit this isn't quite finished and polished yet, but we need it to
prevent users getting confused now KDE auto-detects the system settings. I
plan to have everything fixed up this weekend.  As it stands, most fields
are loaded and displyed and many update, but a few remain to be completed.
Mostly it's just polish.


svn path=/trunk/KDE/kdebase/runtime/; revision=1195805