make socket handling more reliable
make the socket non-blocking, so things don't go awry when a client
freezes. the socket reading will be able to deal with that. the
socket writing function will simply ignore errors - that's not a
problem, as the socket's internal buffer is big enough for our
payloads, so blocking would indicate a serious problem anyway.

set the accept() queue length to 10. no reason to limit it to one.

speculative fix for
CCBUG: 208979
FIXED-IN: 4.5.4

svn path=/branches/KDE/4.5/kdebase/runtime/; revision=1196639