More fixes to locale kcm, and now re-enable saving!
Better solution for checking if config changed, but would be even easier if 
KConfig/KConfigGroup provided an isDirty() call (and a way to open as read 
only would be nice too).

Languages fixed to load/save correctly, but I still can't get the 
KActionSelector widget to work properly, to save tearing my hair out anymore
I'm going to tear it out instead and do it myself.


svn path=/trunk/KDE/kdebase/runtime/; revision=1200855