Make annma happy :-)
My previous idea of a shaded out "en_US" worked in all cases bar one, where the
user is on a system with a system language not en_US and they want to set to
use en_US as their KDE language.  I tried all sorts of other ways of doing this
but they proved too confusing, so it's back to the old way of en_US being a
normal choice with no special handling even though it lets the user do silly
things.  This also means I can switch back to a now fixed KActionSelector.


svn path=/trunk/KDE/kdebase/runtime/; revision=1204930