KCM Locale fix crash when using language other than en_US.
When you pass a KConfig to KLocale it calls setLocale() on the KConfig using
the KLocales current langauge.  If this language is different to the KConfig's
language then a reparseConfiguration() is called to change any localized
settings.  However, if your KCOnfig is enitrely in memory then you have no file
to reparse and you end up with null settings leading to crashes.

Fix makes sure that the KConfig and KLocale always have the same langauge set
before passing in the KConfig, thus making sure there's no force reparse and
the settings remain intact.

BUG: 259790


svn path=/trunk/KDE/kdebase/runtime/; revision=1206247