Redesign the internals of the Resource Merger
The graph in which the statements will be merged is only created if there are statements that should be merged. This solves the issue of having an empty graph if no statements were merged. A custom graph with additional metadata can also be chosen - No longer are we confined to a graph of one type.