Added a ResourceMerger::mergeGraph( uri ) function
This takes the old graph as a parameter and creates a new graph, which has all the properties from both the graph - The new one and the old one. This isn't perfect as a new graph gets created per statement that already exists. I'm going to change this, but I just want to commit my changes.