Added special handling of nie:url to addProperty and setProperty: - Make sure nie:url values are unique - Make sure one resource only has one nie:url value at all times - Setting nie:url on local files will automatically update nfo:fileName, nie:isPartOf   and the nie:url of children if the resources is a file (this is what the MetadataMover   in the filewatchservice used to do - only a much better and faster implementation)
Remark: one could think that the DMS could take over handling of nie:url altogether and
not allow to remove or add that property. The problem with that thought is that not only
local files can have nie:url values. And we cannot handle remote files or web pages
automatically.