Better Graph Merging
The earlier form of graph merging was based on the nao:maintainedBy property. If the old graph, and the new graph had the same maintainer, they were considered the same, and the no merging was performed. The new method is based on the old graph uri.

The new graph that is created as a result of the merge of oldGraph and the additionalMetadata is stored in a hash, so that if another statement contains oldGraph, the merging process is not done again.

We need better type and domain/range checking for the graph metadata!