New createIdentifcationSet( logfile ) method
This method does NOT generate the entire identification set in one go and then save it.
It keeps slowly generating the identificationSet and accordingly saving it. This is better as it consumes less memory. Prior to this patch, with a large Nepomuk repository, the identification set would take over 500 Mb in memory.