Properly fix the memory bloat problem
The problem was that the changeLog and identification set produced could be very large in size. Therefore keeping the entire thing in memory was not a viable option. This patch makes sure that entire ChangeLog is never kept in memory while creating a Backup. ( It is still going to be required when restoring a backup )