Check graphMetadata for Errors
When merging all the resources in the ResourceMerger, the additionalGraphMetadata also needs to be checked for errors - Cardinality, and rdfs:domain/range checks. The new function checks for errors, but doesn't do much if an error is found.

We need better error handling in the ResourceMerger. This code should be reused in the DatamanagementModel to check the additionalGraphMetadta everywhere.