Load translation catalogs of phonon backends after creating a backend
Only recently we noticed that while some backends (gstreamer and xine)
have translatable strings and appropriate po files for KDE l10n, their
catalogs do not actually get loaded. This is a humble try to fix this.

Before createBackend() returns with a valid backend (i.e. a backend was
found and successfully loaded) it will now insert its catalog into the
global KLocale object, which then results in nicely localized strings.

The catalog name is equal to the library name of the backend (as found
in the backend's desktop file). Therefore it is absolutely essential
that the library name matches the pot file created by a backend.

CCMail: phonon-backends@kde.org
CCMail: kde-multimedia@kde.org