Rename backupsync ontology to NRIO
The ontology was still called "backupsync", since that is what it was called when this code was in the playground. It's time to move to the new name - NRIO ( Nepomuk Resource Identification Ontology )

Additionally, since rdf:type is a sub-property of nrio:IdentificationProperty. It is no longer requested separately in the IdentificationSetGenerator