Hide the blank URI creation from the client API.
It always bugged me that users of the client API would have to think
of calling createBlankUri() to create storable SimpleResource instances.
The solution is to use a static counter to always use unique ids for
all resources.
Of course in theory the counter could wrap and then no create unique
ids anymore. But for that to happen one would need to keep a lot of
SimpleResource instances in mem for a long time.
If that is really a problem we could add a set remembering the used
ids.