Added support for network mounts via Solid:NetworkShare.
Now network mounts like nfs or smb mounts are also handled. This means
that file URLs are automatically converted to URLs like
nfs://<HOST>/path.
This required a small change where we store the complete URL of the
mount (not the mount path but the KIO-style URL) instead of just the
UUID of the file system.