Properly handling mapping from <file:/> to <res:/> in storeResources
The extra statements that are added during the mapping ie
    <res> rdf:type nfo:FileDataObject and <res> nie:url <file:/>
should only be added as identification statements if <res> already exists.
If <res> does not exist ( mapped to a blank node ) then those statements
should be added both to identification and to the merger.

This fixes testStoreResources_file3 and testStoreResources_file4