Imported the RemovableMediaModel from the Data Management Service.
The RemovableMediaModel automatically translates absolute file URLs into
and from their relative counterpart. This is done for all removable
storages for which the mount path is not fixed. This includes stuff
like USB keys and optical media, but also network shares like NFS or
Samba. While the latter often have a fixed mount path there are
situations in which that is not the case. Additionally if the mount
is not active the filewatch service will not remove all the existing
files in the index since their URLs are not translated and, thus, the
resources are not found under the mount path. As soon as the path is
mounted the URLs are translated again and the file watch service can
remove resource for the files that do no longer exist.
CCMAIL: nepomuk@kde.org