Let the filewatch service monitor all newly mounted filesystems.
In preparation of removing the removable storage service it is necessary
to let the file watch service monitor removable media itself.
Here we use the same check for devices as we use in the removable media
model in the storage service. It would be better to be able to somehow
share this code.