Add more explanations about why DrKonqi requires certain amount of characters in the bug description.
Tells the user that the required information is proportional to other information,
such as the backtrace or the reproducibility rate. Also, check and tell the user
if he/she has previously selected that they could provide some contextual information.
Additionally add some line breaks to make it easier to read.

REVIEW: 101203
BUG: 268734