Added Nepomuk indexing config to the removable device KCM.
Nepomuk can index files on removable devices. The idea is to ask the
user for each newly mounted device and then remember the choice. Since
this is a configuration similar to the automounting it fits perfectly
into that KCM.
This commit implements the configuration of known devices with respect
to indexing.

Still todo: the Nepomuk configuration is hard coded in many places.
This should be changed to shared code or at least a ConfigXT class.