Added notification to ask the user if a removable device should be indexed.
Added a notification that provides a set of actions when a new removable
device is mounted. The user can choose to index the files on the device
or not or to index files on all newly mounted removable devices.

(This is unfinished: the notification is ugly as it has too many actions
and the config saving is wrong since with UDisks the UDIs do not differ
between devices.)