Do not allow to set values on abstract properties.
In this case abstract properties are defined by the absense of a range.
The only exception we tolerate is nao:identifier for which we assume
an xsd:string range. This is only because nao:identifier has been used
for a long time on Nepomuk.

This change also required some additional ontology statements for the
unit tests.

I also add a unit test written by Vishesh Handa which exposed the
problem with abstract properties since the ClassAndPropertyTree did
convert the values to URIs.