Donot remove all the unwanted entries in one go
Remove them in batches of 20. For some reason if we try to shutdown the strigi
service while it is removing some graph, it just freezes in the local event
loop created by the removeDataByApp job.

Once way to fix this is to forcibly terminate the index scheduler graph on exit
in the ~StrigiService()

If we try to remove more than 20 resources in one go, the DMS times out
occasionally, and that causes massive memory leaks.

CCMAIL: trueg@kde.org