phonon: Remove the remove button.
The remove button was only activated for devices which were not enabled,
but it was not possible to select devices which were not enabled so this
condition was never met. Additionally, the device would only be removed
if the Phonon device was returned in a DBus call which checked to see
if the device was not available. This would always be true as the only way
the button could be valid is if the device was not available.

Thus the DBus call was pointless in the first place and likely should
also be nuked.