Move the backup restoration code from the service into the gui
No one is going to try and restore a backup without the GUI. How
else are you going to resolve conflicts?

The code however was never designed to run in a GUI app, and causes
it to freeze a lot. Proper threading will be required.

For the purpose of getting it to compile and work, I've made both
the Merger and Identifier singletons.