Improve testStoreDataByApplication_folder
This test wasn't entirely correct. It checked for the existance of properties
using ?r nie:isPartOf, when that statement would have been removed.

This test fails - The nao:lastModified is not updated when the the parent
folder resource is removed and nie:isPartOf statement is removed from all its
sub-file resources.

Much to my dissapointment this isn't the miracle test which would make the
strigi service not re-indexed a lot of files on startup.