Add an event loop in the IndexScheduler
Refactor most of the IndexScheduler code so that we have an event loop.
The indexing work has been shifted to a function called doIndexing()
which is called after certain internvals ( based on the indexing speed )
with the help of a QTimer.

This fixes a memory leak, where certain dbus messages would get queued
up, and that would result in the strigi service consuming large amounts
of memory.

This patch is not the ultimate solution as now the strigi service
becomes un-responsive to all dbus calls when it is indexing some files.
Need to investigate more.