Fixed two more special cases in removeDataByApplication.
Only the mtime of resources that were actually changed should be
updated. This means that we need to restrict the query to graphs
maintained by the calling app.

Graphs that are maintained by the calling app but only contain
relations TO the removed resources are now also handled properly.

Both issues are covered in unit tests, too.