Convert QVariants that contain QDbusArguments to the appropriate values.
QtDBus can put any struct in a variant. But it is up to the client to
figure out what is in the variant by looking at the QDBusArgument which
is embedded in the QVariant instead of the actual value.
We already did this for the PropertyHash but forgot to handle the
variant lists.