Donot allow unknown protocols in the DataMangementModel
Change LocalFileState to UriState.

With this patch, any url which is of the form file:/ or any other
protocol that kio supports will be treated as the nie:url automatically.

Unknown protocols like "randomItem:/" results in an error.

This breaks a lot of tests, as many of them use urls like "res:/"