Fix global shortcuts that include symbols produced with shift key
KKeySequenceWidget (used to enter shortcuts) removes shift from the
recorded shortcut if the symbol produced from that key is different
when shift is used. kglobalaccel needs to include shift in the grab in
order to be triggered on this class of shortcuts, and then in the
keypress event handler it also needs to strip the shift again before
checking which shortcut was just triggered.
BUG: 179504
BUG: 197548
BUG: 215030
CCMAIL: kde@michael-jansen.biz
REVIEW: 101520