Make global shortcuts involving shift+tab work
When entering (ctrl,alt,meta)+shift+tab as a new shortcut
KKeySequenceWidget gets "backtab" as the key pressed but records "tab".
kglobalaccel gets the keypress event as "backtab" which is correct but
only triggers shortcuts with "backtab" in them and such shortcuts are
not possible for the user to enter.

This commit makes kglobalaccel look for both tab and backtab when
triggered, same as Qt does. By triggering on both it also avoids
breaking peoples alt+shift+tab on upgrade.

BUG: 174142
BUG: 258467
BUG: 261296
BUG: 274006
CCMAIL: kde@michael-jansen.biz
REVIEW: 101523