Fix some tests
Some of the tests use uris like "res:/A", which are not acceptable unless
they already exist in the Nepomuk repository. Use uris of the form
"nepomuk:/".

Unlike the last commit which changed this, this patch does not use the
"nepomuk" scheme everywhere as in some places, it is first directly added in
the model, and then the dms function are used.

"class:/type*" have been added as RDFS:Classes