Removed any legacy data removal from the filke indexer service.
This task is now done by the data management service in method
removeDataByApplication. All that is left to do is to gather all
resources that have information stored by the file indexer, either in
the old way (via http://www.strigi.org/fields#indexGraphFor) or the new
way (via nao:maintainedBy as created by the DMS).