Fix testStoreResources_lazyCardinalities
If you provide incorrect data, you cannot expect the identification to
be perfect.

If a res with nco:fullname "Superman" already exists and you try
to store a resource and provide nco:fullname "Superman", and
nco:fullname "Clark Kent", then the identification will not be
sucessful, as nco:fullname is an identiying property, and all
identifying properties must either match or not exist.