Check if the graph additionalMetadata is valid before pushing
The graphMetadata will first be converted into a
QMulitHash<QUrl, Soprano::Node>, and then checked for validity. If
it fails, the entire merge operations fails.

This should improve the performance a little bit, as the
additionalMetadata is now being converted into a
QMultiHash<QUrl, Soprano::Node> only once.

This fixes testStoreResources_graphMetadataFail