ResourceMerger : Resolve the nodes in two steps
Initially resolve all the mapped (identified) nodes, then perform
all the max cardinality and domain/range checks. Only after that
should the unmapped nodes be created.

Before this, the nodes were only resolved after the cardinality and
domain range checks. This resulted in the domain range checks
occasionally failing if the required type was in the store.

This fixes : testStoreResources_correctDomainInStore2
             testStoreResources_correctRangeInStore2

Resolving a node => changing it's uri to the identified resource