Do not initialize the ResourceManager's mainModel
The ResourceIdentifier used to call Sync::ResourceIdentifier which
used to set the model to ResourceManager::instance()->mainModel(),
and then a couple of lines later, we would explicitly set the
ResourceIdentifer's model via setModel( )

Nepomuk::ResourceManager::instance()->mainModel() should never
be called in the storage service. We now set the model directly in
the ResourceIdentifier's constructor.