Allow addTriplet to add information about the main resource.
So far we worked under the assumption that information about the main
resource would only be added through the addValue methods. That is not
always the case. There is for example the odf analyzer which does use
addTriplet instead. Thus, we need to check the subject for the path of
the main resource before continuing to map sub-resources.