Remove monthly auto backups from the GUI
Monthly auto backup was removed in 4.6 with
http://websvn.kde.org/?view=revision&revision=1215783 with this log:
Removed monthly auto backups in 4.6. It was missing the date configuration
and does not seem to be implemented in the service anyway.
Not implemented in 4.7 or master, so remove it from the gui
REVIEW:102195
CCBUG:263530
(cherry picked from commit c02dc754657552bf35c2e2380701589d9d97e792)