Handle special case in removeDataByApplication.
When a call to removeDataByApplication removes the entire resource but
there is still an incoming relation created by another application
then we keep the metadata of the resource. But only if the resource has
a nie:url. Without a nie:url the left over metadata is pointless but
with a nie:url the resource can still be identified even outside of
Nepomuk.
This fix is important for any indexer like the file indexer or the
Akonadi indexer since they use removeDataByApplication do update
resources. Without this fix incoming relations would be deleted and
the resource URI would change.

The same fix still has to be implemented for the other variant of
removeDataByApplication. It is, however, by far not as urgent since
it is not used anywhere yet.