Print error message to stdout in main.cpp and use valid return codes.
Previously the indexer would always return 0 even in the error case.
Now it returns 1 in case indexing failed.

The error message is actually useful and should be separated from the
debugging output. We can even use it later on to log the files which
could not be indexed from the service.