Inserts the program version in more cases in bug reports.
As it turns out the release of KDE 4.7.0 is called "4.7.00"
so with two 0s, as was also true for 4.6.0.
Yet on bugs.kde.org there often is just e.g. version "4.7.0"
available, thus DrKonqi does not insert the version in the version
field.
Now DrKonqi treats "4.7.0" and "4.7.00" as equal when determaining
if the verion should be entered in the version field.

REVIEW:102479