Fixed KHelpCenter Font Scaling (bug 243082)
When user clicked a button to change font size the entire document was
zoomed, instead of just the font size. This meant that the text would
run past the right edge and user had to scroll horizontaly to read the
whole line. After this patch only the font size changes.

Thanks to Jonathan Ryshpan for bringing this bug to my attention.
BUG: 243082
FIXED-IN: 4.7.2
CCMAIL: jonrysh@pacbell.net
REVIEW: 102528