Fix jobs some times showing "empty"
Basically since we use all async dbus methos we can't be sure
that once the contact is added before any variable is set. What
this patch does is send all the cached variables to the new added
contact so we can be sure that it contains all the up to date data.

REVIEW: 102525