Allow to configure email indexing in Nepomuk KCM.
This adds an option to the Nepomuk KCM to disable email full text indexing.
It's somewhat comparable to file system indexing in terms of resource
usage, so there has been demand for such an option and this seems to be
the natural place to put it.

This goes together with a number of patches to kdepim that make KMail and
the email indexing agent honor this setting, which have been committed
already.

REVIEW: 102440