Use new versions of the default exclude filter list.
This is achieved by using the new default exclude filter list version.
Since the KCM does not know about the version field yet (we need to
make the config class public and use it in the KCM) we re-write the
exclude filter list entry whenever it is requested and the version
does not match.
There is one problem here which I do not know how to solve yet:
if the user removed some of the default exclude filters they will be
re-added once the default list version changes. The simplest solution
would be to not let the user remove filters from the default filter
list and not even show them in the KCM.

BUG: 283562
FIXED-IN: 4.7.3