Make the phonon kcm backend build optional
CCMAIL: sitter@kde.org

The current issue is that we do not have a proper libcanberra version on
Harmattan, as in: the current version is 0.12, and the upstream debian version
is 0.28 for instance. The old version installation breaks the build of the
kde-runtime project.

Unfortunately, platform packages cannot be replaced by builds and versions
according to the Harmattan Security Platform policies.

There are three approaches for solving this issue:
1) Asking for update from the Harmattan SDK maintainers. I have tried to build a
new package and provide it internally to the Harmattan SDK Maintainers, but it
did not work out. The answer was "too much risk".

2) Install the new version into a new location on the system by using different
prefix path, and then use -DCMAKE_PREFIX_PATH and similar include/library path
hackarounds, but that goes to a messed up system

3) Make the phonon kcm backend build optional and we can skip it for now on
Harmattan. There will hopefully be other backend that will be sorted out for
phonon operations, and we can be happy.

I went for the third option since that seems to be the most reasonable to me
since it preserves the functionality and people can get rid of that backend
where it causes issues. Feel free to revert the patch if it is not acceptable,
and choosing this way is the limitation of my knowledge about kde-runtime and
phonon.

Testing: I have tested the patch on Desktop and Harmattan. It works as expected,
and I could not find any regressions yet.