Backport from nepomuk-core:
Fixed StrigiServiceConfig::shouldBeIndexed for exclude filters.

We did not take one issue into account before: imagine we have an
exclude filter "build" and a folder $HOME/x/build/y/. This folder
should not be indexed. This was not handled before.
By checking each component of a path starting from the include folder
this case is now handled, too.

One problem remains: the IndexCleaner does not take this issue into
account yet. This means it will not cleanup all required entries yet.

BUG: 279351
FIXED-IN: 4.7.3