Handle kext:altUrl in the file watch service.
kext:altUrl is used to store URLs of indirect symlinks. These need to
be updated along with a file move or removal, too.
If a symlink to a folder is removed we need to subsequently remove all
kex:altUrl properties which have a value URL starting with that
particular URL.
If a link is renamed we only need to update the nie:url. The DMS does
the rest as it does for nie:urls as well.