Fix testStoreResources_objectExistsIdentification
If we add statements of the form
    _:a <someProp> <nepomuk:/uri-that-already-exists>

without including <nepomuk:/uri-that-already-exists> in the SimpleResource
we get an error that '<nepomuk:/uri-that-already-exists>' doesn't exist.

The fix isn't ideal as this could also be done in the Identification phase,
but this is just simpler. Plus, I landed up messing up too many tests
when I tried to fix it in the identification phase.