Check the modification time before reindexing files
The modification time was always checked, but only when a directory
was sent for indexing. The analyzeDir function would check for files
as well. This worked fine the cases where the Indexer would routinely
check and re-index files.

It however didn't work so well in the cases where the file watcher
told the indexer to reindex a file cause of a inotify event. Specifically
CloseOnWrite event, which are emmited when a file opened with "r" mode
is closed. Even if modifications aren't made.

Applications like Amarok and KTorrent routinely open files in "rw" mode
even when they don't have to make changes. Those files used to get
re-indexed.

REVIEW: 103794