fix left side of shadow/item in tasks svg
good news everyone!! from what i can tell the annoying vertical lines,
as well as mismatching gradients in the tasks svgz's are all fixed now.
the Real Fix would be to make them vector, they shouldn't be raster, and
i'm sure it can be avoided/fixed. but i'm a hacker, not so much an
artist ;)